[ADD] JSON and HTTP Benchmarks (AI generated, Im not going to write 1.6k lines of handmade benchmark and deal with CMake hell). [MODIFY] Some optimizations. [MODIFY] CMAKE options to build examples, tests and benchmark.
Os benchmarks parecem meio enviesados (principalmente HTTP) mas blz, dps arrumo. Cansei bastante agr...

+++ include/Thoth/NJson/Definitions.hpp
@@ -23,7 +23,7 @@ namespace Thoth::NJson {
 
     using Null   = std::monostate;                   // null
     using String = Dsa::Cow<StringRef, std::string>; // string
-    using Number = long double;                      // number
+    using Number = double;                           // number
     using Bool   = bool;                             // bool
     using Object = std::unique_ptr<JsonObject>;      // {Object}
     using Array  = std::vector<Json>;                // [Array]
